only admin and moderator can dismiss reports

+++ app/Http/Controllers/ReportController.php
@@ -34,6 +34,9 @@ class ReportController extends Controller
     }
 
     public function dismiss(Request $request){
+        if(!Auth::user()->moderator && !Auth::user()->administrator)
+            abort(404);
+                
         $found = Report::where('id_post',$request->id_post)->delete();
         
         return redirect()->back();
